refactor(middleware): move path tracking to dedicated path.go file
